Extract ShelfContentCard + init. Discover section
